stemstudio/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
demucs==4.0.1
deepfilternet==0.5.6
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic==2.7.0
